Card design and order
The cards are now in a landscape design, and both the icon are text are arranged horizontally instead of top-down. The only problem is that they are all on the left side of the card and overlapping one another. Secondly, implemented displaying names in order of who has the highest number of 'buck_amount'.